N Queens problem
The eight queens puzzle is the problem of placing eight chess queens on an 8×8 chessboard so that no two queens threaten each other.
More general N-Queens puzzle ( placing n queens on an n×n chessboard, where solutions exist for all natural numbers n with the exception of n=2 and n=3) is a classic computer science problem. In fact, it’s much older than discipline of computer science. It is usually used as a problem to introduce students to backtracking algorithms in computer science.